updates per review comments
